+++ src/native/NodeConfig.h
@@ -8,11 +8,11 @@ struct NodeConfig {
     int pixelCount;
     std::array<int, 3> cubeCoordinates;
     std::array<int, 2> ringCoordinates;
-    std::array<int, 2> cartesian2dCoordinates;
+    std::array<float, 2> cartesian2dCoordinates;
 
 public:
     NodeConfig(int id, int count, const std::array<int, 2> ringCoords,
-                           const std::array<int, 3> cubeCoords, const std::array<int, 2> cartesianCoords)
+                           const std::array<int, 3> cubeCoords, const std::array<float, 2> cartesianCoords)
         : nodeId(id), pixelCount(count), ringCoordinates(ringCoords),
           cubeCoordinates(cubeCoords), cartesian2dCoordinates(cartesianCoords)
     {
@@ -22,7 +22,7 @@ public:
     int getPixelCount() const;
     const std::array<int, 2> getRingCoordinates() const;
     const std::array<int, 3> getCubeCoordinates() const;
-    const std::array<int, 2> getCartesian2dCoordinates() const;
+    const std::array<float, 2> getCartesian2dCoordinates() const;
 };
 
 #endif // NODECONFIG_H
